requirements.txt
--index-url https://download.pytorch.org/whl/cu121
torch==2.3.1
torchvision==0.18.1
numpy==1.26.3
tqdm==4.64.1
kornia==0.7.3
matplotlib==3.9.1
notebook==7.2.1
Pillow==10.2.0
gdown==5.2.0
nibabel==5.2.1